+++ spec/components/attachment_spec.rb
@@ -269,19 +269,18 @@ describe "Attachment", type: :view do
     assert_select '.gem-c-attachment:not([class*="govuk-!-margin-bottom-"])'
   end
 
-  it "adds GA4 tracking on HTML attachment links" do
+  it "includes GA4 tracking on attachment links by default" do
     render_component(
       attachment: {
         title: "Test",
-        url: "https://assets.publishing.service.gov.uk/government/uploads/system/uploads/attachment_data/file/687542/February_2018_Consular_MI.csv",
+        url: "https://example.com",
         filename: "test",
         type: "html",
       },
     )
 
-    assert_select "h2 a[data-module=ga4-link-tracker]"
-    assert_select ".gem-c-attachment__thumbnail a[data-module=ga4-link-tracker]"
-    assert_select "h2 a[data-ga4-link='{\"event_name\":\"navigation\",\"type\":\"attachment\"}']"
-    assert_select ".gem-c-attachment__thumbnail a[data-ga4-link='{\"event_name\":\"navigation\",\"type\":\"attachment\"}']"
+    assert_select "section[data-module=ga4-link-tracker]"
+    assert_select "section[data-ga4-link='{\"event_name\":\"navigation\",\"type\":\"attachment\"}']"
+    assert_select "section[data-ga4-track-links-only]"
   end
 end